split setup and questions into 2 files
